requirements.txt
# Основные зависимости
python-dotenv==1.0.0
SpeechRecognition==3.10.0
openai==0.28.1
pyttsx3==2.90
pyautogui==0.9.54
keyboard==0.13.5
pyperclip==1.8.2

# Для распознавания речи с помощью Whisper
openai-whisper==20231117
torch==2.0.1
numpy==1.24.4

# Дополнительные зависимости для улучшения качества звука
PyAudio==0.2.13 